scripts/pyproject.toml
[tool.poetry]
name = "support scripts"
version = "0.1.0"
description = ""
authors = ["prakashsellathurai <prakashsellathurai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
feedparser = "^6.0.8"
pyaml = "^21.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
